simplified SDK structure
