--- packages/sdk/src/base/index.ts
@@ -1,4 +1,0 @@
-export * from './storage';
-
-export * from './auth';
-export * from './items';
