+++ packages/sdk/src/handlers/permissions.ts
@@ -2,7 +2,7 @@
  * Permissions handler
  */
 
-import { ItemsHandler } from '../base/items';
+import { ItemsHandler } from '../items';
 import { Transport } from '../transport';
 import { PermissionType, DefaultType } from '../types';
