+++ packages/sdk/src/handlers/presets.ts
@@ -2,7 +2,7 @@
  * Presets handler
  */
 
-import { ItemsHandler } from '../base/items';
+import { ItemsHandler } from '../items';
 import { Transport } from '../transport';
 import { PresetType, DefaultType } from '../types';
