+++ packages/sdk/src/handlers/revisions.ts
@@ -2,7 +2,7 @@
  * Revisions handler
  */
 
-import { ItemsHandler } from '../base/items';
+import { ItemsHandler } from '../items';
 import { Transport } from '../transport';
 import { RevisionType, DefaultType } from '../types';
