+++ packages/sdk/src/handlers/roles.ts
@@ -2,7 +2,7 @@
  * Roles handler
  */
 
-import { ItemsHandler } from '../base/items';
+import { ItemsHandler } from '../items';
 import { Transport } from '../transport';
 import { RoleType, DefaultType } from '../types';
