+++ packages/sdk/src/index.ts
@@ -2,13 +2,8 @@
 export * from './auth';
 export * from './directus';
 export * from './handlers';
-export * from './items';
-export * from './singleton';
 export * from './storage';
 export * from './transport';
-
-// Implementations
-export * from './base';
 
 // Types
 export * from './types';
